Split internal/ui/ui.go into cohesive files
The 1375-line ui.go mixed six concerns. Split it by responsibility with
no behavior change: style.go (color/hyperlink helpers), status.go
(narration + Term* summaries), spinner.go (spinner writer + lifecycle),
progress.go (worker-slot API + debug tracing), text.go (width/truncation
utilities). ui.go keeps the struct, constructors, mode detection, and the
emit/log plumbing.

Pure code movement: the UI golden characterization tests and the existing
spinner tests pass unchanged, proving the output surface is byte-identical.
